refactor: split io path runtime emitters
