+++ src/codegen/runtime/io/mod.rs
@@ -1,4 +1,7 @@
+mod basename;
 mod cstr;
+mod dirname;
+mod dirname_levels;
 mod feof;
 mod fgetcsv;
 mod fgets;
@@ -13,8 +16,8 @@ mod fs;
 mod getcwd;
 mod glob;
 mod modify;
-mod path;
-mod pathinfo;
+mod pathinfo_array;
+mod pathinfo_str;
 mod realpath;
 mod scandir;
 mod stat;
@@ -22,7 +25,10 @@ mod stat_array;
 mod stat_ext;
 mod tempnam;
 
+pub(crate) use basename::emit_basename;
 pub(crate) use cstr::emit_cstr;
+pub(crate) use dirname::emit_dirname;
+pub(crate) use dirname_levels::emit_dirname_levels;
 pub(crate) use feof::emit_feof;
 pub(crate) use fgetcsv::emit_fgetcsv;
 pub(crate) use fgets::emit_fgets;
@@ -37,11 +43,8 @@ pub(crate) use fs::emit_fs;
 pub(crate) use getcwd::emit_getcwd;
 pub(crate) use glob::emit_glob;
 pub(crate) use modify::emit_modify;
-pub(crate) use path::emit_basename;
-pub(crate) use path::emit_dirname;
-pub(crate) use path::emit_dirname_levels;
-pub(crate) use pathinfo::emit_pathinfo_array;
-pub(crate) use pathinfo::emit_pathinfo_str;
+pub(crate) use pathinfo_array::emit_pathinfo_array;
+pub(crate) use pathinfo_str::emit_pathinfo_str;
 pub(crate) use realpath::emit_realpath;
 pub(crate) use scandir::emit_scandir;
 pub(crate) use stat::emit_stat;
